// src/pages/company/library/data.d.ts
export type LibraryItem = {
  key: number;
  title: string;
  author: string;
  desc: string;
  date: number;
};

export type Pagination = {
  total: number;
  pageSize: number;
  current: number;
};

export type Params = {
  status?: string;
  name?: string;
  desc?: string;
  key?: number;
  pageSize?: number;
  currentPage?: number;
  filter?: Record<string, any[]>;
  sorter?: Record<string, any>;
};